worker/requirements.txt
redis==5.0.1
rq==1.15.1
playwright==1.40.0
psycopg2-binary==2.9.9
requests==2.31.0
tldextract==5.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
python-Levenshtein==0.23.0
